+++ docs/js/menu-wc.js
@@ -1,0 +1,240 @@
+'use strict';
+
+customElements.define('compodoc-menu', class extends HTMLElement {
+    constructor() {
+        super();
+        this.isNormalMode = this.getAttribute('mode') === 'normal';
+    }
+
+    connectedCallback() {
+        this.render(this.isNormalMode);
+    }
+
+    render(isNormalMode) {
+        let tp = lithtml.html(`<nav>
+    <ul class="list">
+        <li class="title">
+            <a href="index.html" data-type="index-link">spaier-ng-recaptcha documentation</a>
+        </li>
+        <li class="divider"></li>
+        ${ isNormalMode ? `<div id="book-search-input" role="search">
+    <input type="text" placeholder="Type to search">
+</div>
+` : '' }
+        <li class="chapter">
+            <a data-type="chapter-link" href="index.html"><span class="icon ion-ios-home"></span>Getting started</a>
+            <ul class="links">
+                    <li class="link">
+                        <a href="overview.html" data-type="chapter-link">
+                            <span class="icon ion-ios-keypad"></span>Overview
+                        </a>
+                    </li>
+                    <li class="link">
+                        <a href="index.html" data-type="chapter-link">
+                            <span class="icon ion-ios-paper"></span>README
+                        </a>
+                    </li>
+                    <li class="link">
+                            <a href="changelog.html"
+                        data-type="chapter-link">
+                            <span class="icon ion-ios-paper"></span>CHANGELOG
+                        </a>
+                    </li>
+                    <li class="link">
+                            <a href="contributing.html"
+                        data-type="chapter-link">
+                            <span class="icon ion-ios-paper"></span>CONTRIBUTING
+                        </a>
+                    </li>
+                    <li class="link">
+                            <a href="license.html"
+                        data-type="chapter-link">
+                            <span class="icon ion-ios-paper"></span>LICENSE
+                        </a>
+                    </li>
+            </ul>
+        </li>
+        <li class="chapter modules">
+            <a data-type="chapter-link" href="modules.html">
+                <div class="menu-toggler linked" data-toggle="collapse"
+                    ${ isNormalMode ? 'data-target="#modules-links"' : 'data-target="#xs-modules-links"' }>
+                    <span class="icon ion-ios-archive"></span>
+                    <span class="link-name">Modules</span>
+                    <span class="icon ion-ios-arrow-down"></span>
+                </div>
+            </a>
+            <ul class="links collapse"
+            ${ isNormalMode ? 'id="modules-links"' : 'id="xs-modules-links"' }>
+                    <li class="link">
+                        <a href="modules/AppModule.html" data-type="entity-link">AppModule</a>
+                            <li class="chapter inner">
+                                <div class="simple menu-toggler" data-toggle="collapse"
+                                    ${ isNormalMode ? 'data-target="#components-links-module-AppModule-1cd8d066731d25fbb80bf0a28f1431cc"' : 'data-target="#xs-components-links-module-AppModule-1cd8d066731d25fbb80bf0a28f1431cc"' }>
+                                    <span class="icon ion-md-cog"></span>
+                                    <span>Components</span>
+                                    <span class="icon ion-ios-arrow-down"></span>
+                                </div>
+                                <ul class="links collapse"
+                                    ${ isNormalMode ? 'id="components-links-module-AppModule-1cd8d066731d25fbb80bf0a28f1431cc"' : 'id="xs-components-links-module-AppModule-1cd8d066731d25fbb80bf0a28f1431cc"' }>
+                                        <li class="link">
+                                            <a href="components/AppComponent.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">AppComponent</a>
+                                        </li>
+                                        <li class="link">
+                                            <a href="components/DirectiveComponent.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">DirectiveComponent</a>
+                                        </li>
+                                        <li class="link">
+                                            <a href="components/ReactiveComponent.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">ReactiveComponent</a>
+                                        </li>
+                                        <li class="link">
+                                            <a href="components/TemplateComponent.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">TemplateComponent</a>
+                                        </li>
+                                </ul>
+                            </li>
+                    </li>
+                    <li class="link">
+                        <a href="modules/AppRoutingModule.html" data-type="entity-link">AppRoutingModule</a>
+                    </li>
+                    <li class="link">
+                        <a href="modules/MaterialModule.html" data-type="entity-link">MaterialModule</a>
+                    </li>
+                    <li class="link">
+                        <a href="modules/RecaptchaFormsModule.html" data-type="entity-link">RecaptchaFormsModule</a>
+                            <li class="chapter inner">
+                                <div class="simple menu-toggler" data-toggle="collapse"
+                                    ${ isNormalMode ? 'data-target="#directives-links-module-RecaptchaFormsModule-22a994ae29b3d56cbabaa3037de41afe"' : 'data-target="#xs-directives-links-module-RecaptchaFormsModule-22a994ae29b3d56cbabaa3037de41afe"' }>
+                                    <span class="icon ion-md-code-working"></span>
+                                    <span>Directives</span>
+                                    <span class="icon ion-ios-arrow-down"></span>
+                                </div>
+                                <ul class="links collapse"
+                                    ${ isNormalMode ? 'id="directives-links-module-RecaptchaFormsModule-22a994ae29b3d56cbabaa3037de41afe"' : 'id="xs-directives-links-module-RecaptchaFormsModule-22a994ae29b3d56cbabaa3037de41afe"' }>
+                                        <li class="link">
+                                            <a href="directives/RecaptchaDirective.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">RecaptchaDirective</a>
+                                        </li>
+                                </ul>
+                            </li>
+                    </li>
+                    <li class="link">
+                        <a href="modules/RecaptchaLoaderModule.html" data-type="entity-link">RecaptchaLoaderModule</a>
+                            <li class="chapter inner">
+                                <div class="simple menu-toggler" data-toggle="collapse"
+                                    ${ isNormalMode ? 'data-target="#injectables-links-module-RecaptchaLoaderModule-df529a3ff4814faa5a0bd6860794ac6c"' : 'data-target="#xs-injectables-links-module-RecaptchaLoaderModule-df529a3ff4814faa5a0bd6860794ac6c"' }>
+                                    <span class="icon ion-md-arrow-round-down"></span>
+                                    <span>Injectables</span>
+                                    <span class="icon ion-ios-arrow-down"></span>
+                                </div>
+                                <ul class="links collapse"
+                                    ${ isNormalMode ? 'id="injectables-links-module-RecaptchaLoaderModule-df529a3ff4814faa5a0bd6860794ac6c"' : 'id="xs-injectables-links-module-RecaptchaLoaderModule-df529a3ff4814faa5a0bd6860794ac6c"' }>
+                                        <li class="link">
+                                            <a href="injectables/RecaptchaService.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules"}>RecaptchaService</a>
+                                        </li>
+                                </ul>
+                            </li>
+                    </li>
+                    <li class="link">
+                        <a href="modules/RecaptchaModule.html" data-type="entity-link">RecaptchaModule</a>
+                            <li class="chapter inner">
+                                <div class="simple menu-toggler" data-toggle="collapse"
+                                    ${ isNormalMode ? 'data-target="#directives-links-module-RecaptchaModule-8563b04f8c6dfcbea372276188701923"' : 'data-target="#xs-directives-links-module-RecaptchaModule-8563b04f8c6dfcbea372276188701923"' }>
+                                    <span class="icon ion-md-code-working"></span>
+                                    <span>Directives</span>
+                                    <span class="icon ion-ios-arrow-down"></span>
+                                </div>
+                                <ul class="links collapse"
+                                    ${ isNormalMode ? 'id="directives-links-module-RecaptchaModule-8563b04f8c6dfcbea372276188701923"' : 'id="xs-directives-links-module-RecaptchaModule-8563b04f8c6dfcbea372276188701923"' }>
+                                        <li class="link">
+                                            <a href="directives/RecaptchaDirective.html" data-type="entity-link" data-context="sub-entity" data-context-id="modules">RecaptchaDirective</a>
+                                        </li>
+                                </ul>
+                            </li>
+                    </li>
+            </ul>
+        </li>
+                <li class="chapter">
+                    <div class="simple menu-toggler" data-toggle="collapse"
+                    ${ isNormalMode ? 'data-target="#directives-links"' : 'data-target="#xs-directives-links"' }>
+                        <span class="icon ion-md-code-working"></span>
+                        <span>Directives</span>
+                        <span class="icon ion-ios-arrow-down"></span>
+                    </div>
+                    <ul class="links collapse"
+                    ${ isNormalMode ? 'id="directives-links"' : 'id="xs-directives-links"' }>
+                            <li class="link">
+                                <a href="directives/RecaptchaValueAccessorDirective.html" data-type="entity-link">RecaptchaValueAccessorDirective</a>
+                            </li>
+                    </ul>
+                </li>
+                <li class="chapter">
+                    <div class="simple menu-toggler" data-toggle="collapse"
+                        ${ isNormalMode ? 'data-target="#injectables-links"' : 'data-target="#xs-injectables-links"' }>
+                        <span class="icon ion-md-arrow-round-down"></span>
+                        <span>Injectables</span>
+                        <span class="icon ion-ios-arrow-down"></span>
+                    </div>
+                    <ul class="links collapse"
+                    ${ isNormalMode ? 'id="injectables-links"' : 'id="xs-injectables-links"' }>
+                            <li class="link">
+                                <a href="injectables/RecaptchaOnloadService.html" data-type="entity-link">RecaptchaOnloadService</a>
+                            </li>
+                            <li class="link">
+                                <a href="injectables/RecaptchaService.html" data-type="entity-link">RecaptchaService</a>
+                            </li>
+                    </ul>
+                </li>
+        <li class="chapter">
+            <div class="simple menu-toggler" data-toggle="collapse"
+                ${ isNormalMode ? 'data-target="#interfaces-links"' : 'data-target="#xs-interfaces-links"' }>
+                <span class="icon ion-md-information-circle-outline"></span>
+                <span>Interfaces</span>
+                <span class="icon ion-ios-arrow-down"></span>
+            </div>
+            <ul class="links collapse"
+            ${ isNormalMode ? ' id="interfaces-links"' : 'id="xs-interfaces-links"' }>
+                    <li class="link">
+                        <a href="interfaces/Recaptcha.html" data-type="entity-link">Recaptcha</a>
+                    </li>
+                    <li class="link">
+                        <a href="interfaces/RecaptchaExecuteParameters.html" data-type="entity-link">RecaptchaExecuteParameters</a>
+                    </li>
+                    <li class="link">
+                        <a href="interfaces/RecaptchaParameters.html" data-type="entity-link">RecaptchaParameters</a>
+                    </li>
+            </ul>
+        </li>
+        <li class="chapter">
+            <div class="simple menu-toggler" data-toggle="collapse"
+            ${ isNormalMode ? 'data-target="#miscellaneous-links"' : 'data-target="#xs-miscellaneous-links"' }>
+                <span class="icon ion-ios-cube"></span>
+                <span>Miscellaneous</span>
+                <span class="icon ion-ios-arrow-down"></span>
+            </div>
+            <ul class="links collapse"
+            ${ isNormalMode ? 'id="miscellaneous-links"' : 'id="xs-miscellaneous-links"' }>
+                    <li class="link">
+                      <a href="miscellaneous/enumerations.html" data-type="entity-link">Enums</a>
+                    </li>
+                    <li class="link">
+                      <a href="miscellaneous/functions.html" data-type="entity-link">Functions</a>
+                    </li>
+                    <li class="link">
+                      <a href="miscellaneous/variables.html" data-type="entity-link">Variables</a>
+                    </li>
+            </ul>
+        </li>
+            <li class="chapter">
+                <a data-type="chapter-link" href="routes.html"><span class="icon ion-ios-git-branch"></span>Routes</a>
+            </li>
+        <li class="chapter">
+            <a data-type="chapter-link" href="coverage.html"><span class="icon ion-ios-stats"></span>Documentation coverage</a>
+        </li>
+        <li class="divider"></li>
+        <li class="copyright">
+                Documentation generated using <a href="https://compodoc.app/" target="_blank">
+                            <img data-src="images/compodoc-vectorise.svg" class="img-responsive" data-type="compodoc-logo">
+                </a>
+        </li>
+    </ul>
+</nav>`);
+        this.innerHTML = tp.strings;
+    }
+});